+++ modules/serverpod_auth/serverpod_auth_google_flutter/lib/serverpod_auth_google_flutter.dart
@@ -1,0 +1,5 @@
+library serverpod_auth_google_flutter;
+
+export 'src/auth.dart';
+export 'src/widgets/signin_button.dart';
+export 'package:serverpod_auth_shared_flutter/serverpod_auth_shared_flutter.dart';