+++ modules/serverpod_auth/serverpod_auth_server/lib/src/endpoints/user_endpoint.dart
@@ -6,8 +6,13 @@ import 'package:serverpod/serverpod.dart';
 import '../generated/protocol.dart';
 
 class UserEndpoint extends Endpoint {
+  Future<bool> isSignedIn(Session session) async {
+    var userId = await session.auth.authenticatedUserId;
+    return userId != null;
+  }
+
   Future<UserInfo?> getAuthenticatedUserInfo(Session session) async {
-    var userId = await session.authenticatedUserId;
+    var userId = await session.auth.authenticatedUserId;
     if (userId == null)
       return null;