+++ modules/serverpod_auth/serverpod_auth_shared_flutter/lib/serverpod_auth_shared_flutter.dart
@@ -1,0 +1,3 @@
+library serverpod_auth_shared_flutter;
+
+export 'src/authentication_key_manager.dart';