+++ packages/serverpod_client/lib/src/auth_key_manager.dart
@@ -1,8 +1,8 @@
 /// Manages keys for authentication with the server.
 abstract class AuthenticationKeyManager {
   /// Retrieves an authentication key.
-  Future<String> get();
+  Future<String?> get();
 
   /// Saves an authentication key retrieved by the server.
-  Future<Null> put(String key);
+  Future<void> put(String key);
 }