+++ packages/serverpod_service_client/lib/src/protocol/protocol.dart
@@ -22,6 +22,7 @@ import 'log_level.dart';
 import 'session_log_entry.dart';
 import 'server_health_result.dart';
 import 'session_log_result.dart';
+import 'auth_key.dart';
 import 'readwrite_test.dart';
 
 export 'cache_info.dart';
@@ -39,6 +40,7 @@ export 'log_level.dart';
 export 'session_log_entry.dart';
 export 'server_health_result.dart';
 export 'session_log_result.dart';
+export 'auth_key.dart';
 export 'readwrite_test.dart';
 export 'client.dart';
 
@@ -68,6 +70,7 @@ class Protocol extends SerializationManager {
     constructors['SessionLogEntry'] = (Map<String, dynamic> serialization) => SessionLogEntry.fromSerialization(serialization);
     constructors['ServerHealthResult'] = (Map<String, dynamic> serialization) => ServerHealthResult.fromSerialization(serialization);
     constructors['SessionLogResult'] = (Map<String, dynamic> serialization) => SessionLogResult.fromSerialization(serialization);
+    constructors['AuthKey'] = (Map<String, dynamic> serialization) => AuthKey.fromSerialization(serialization);
     constructors['ReadWriteTestEntry'] = (Map<String, dynamic> serialization) => ReadWriteTestEntry.fromSerialization(serialization);
   }
 }