+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Support for static file directories in Relic.
 - Adds logos (psd and pngs).
+- Adds example project.
+- Initial version of authentication module.
 
 ## 0.8.6